working session stats. (someofem)
i fixed my issue with the unmarshalling that the generic
makeRequest was doing by passing the work onto calling
functions. i feel this is a fair choice because it keeps
the ownership of "knowledge" of a response type and removes
clutter from a singular Response type. Args is now a RawMessage
and can be unmarshalled appropriately by an "intelligent" caller.